+++ lib/hubbado/log/notify_rollbar.rb
@@ -17,6 +17,8 @@ module Hubbado
         @notifier = notifier
       end
 
+      def traces?(severity, _tags = nil) = LEVELS.key?(severity.to_sym)
+
       def log(subject, severity, message, data = nil, stacktrace = nil, _tags = nil)
         level = LEVELS[severity.to_sym]
         return if level.nil?