+++ lib/hubbado/log/rails_logger.rb
@@ -14,6 +14,8 @@ module Hubbado
         @rails_logger = rails_logger
       end
 
+      def traces?(severity, tags = nil) = Display.shows?(severity, tags)
+
       def log(subject, severity, message, data = nil, stacktrace = nil, tags = nil)
         return unless Display.shows?(severity, tags)